refactor: move inclusion-resolver to repository package
