+++ packages/repository/src/relations/has-many/has-many.helpers.ts
@@ -7,7 +7,7 @@ import * as debugFactory from 'debug';
 import {camelCase} from 'lodash';
 import {InvalidRelationError} from '../../errors';
 import {isTypeResolver} from '../../type-resolver';
-import {HasManyDefinition} from '../relation.types';
+import {HasManyDefinition, RelationType} from '../relation.types';
 
 const debug = debugFactory('loopback:repository:has-many-helpers');
 
@@ -27,6 +27,11 @@ export type HasManyResolvedDefinition = HasManyDefinition & {keyTo: string};
 export function resolveHasManyMetadata(
   relationMeta: HasManyDefinition,
 ): HasManyResolvedDefinition {
+  if ((relationMeta.type as RelationType) !== RelationType.hasMany) {
+    const reason = 'relation type must be HasMany';
+    throw new InvalidRelationError(reason, relationMeta);
+  }
+
   if (!isTypeResolver(relationMeta.target)) {
     const reason = 'target must be a type resolver';
     throw new InvalidRelationError(reason, relationMeta);
@@ -58,7 +63,9 @@ export function resolveHasManyMetadata(
   const hasDefaultFkProperty = targetModelProperties[defaultFkName];
 
   if (!hasDefaultFkProperty) {
-    const reason = `target model ${targetModel.name} is missing definition of foreign key ${defaultFkName}`;
+    const reason = `target model ${
+      targetModel.name
+    } is missing definition of foreign key ${defaultFkName}`;
     throw new InvalidRelationError(reason, relationMeta);
   }
