+++ packages/repository/src/relations/has-many/index.ts
@@ -3,6 +3,7 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-export * from './has-many.decorator';
-export * from './has-many.repository';
 export * from './has-many-repository.factory';
+export * from './has-many.decorator';
+export * from './has-many.inclusion-resolver';
+export * from './has-many.repository';
